backend/requirements.txt
black==24.3.0
certifi==2023.11.17
charset-normalizer==3.3.2
celery==5.3.4
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
idna==3.6
mypy-extensions==1.0.0
nltk==3.8.1
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary>=2.8.6
pyodbc==5.0.1
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.7
sqlparse==0.4.4
tk==0.1.0
torch==2.3.1
transformers==4.42.3
typing_extensions==4.8.0
urllib3==2.1.0
